address starting a new session while old dinit is still dying
Imagine the case when a session is terminated, old dinit is sent
SIGTERM, but before the old process could be reaped and its dirfd
could be cleared and so on, a new session is started. We do not
want to start the new dinit right away in those cases, instead
we want to wait for the old one to die first.

Also create the session dir and possibly rundir as a part of
dinit_start. This is because doing it during session creation
may conflict with the old dinit being terminated (as only after
its termination has finished, we want to remove the old dir and
discard the old descriptor, so by the time we are creating a new
one we want to be sure it is clear)