attempt to reuse rundir if managing it
When managing rundir, the user has made it clear that nothing else
is trying to handle it. That means if there is rundir left for that
user, it could be a previous potentially crashed attempt, or a
manually created directory. As long as it's actually a directory,
just ensure it has correct permissions and move on.

Another matter is if this is *not* a directory, then we should
absolutely fail.