pam: add auth and drop loginuid, make turnstile required
It seems some PAM setups will fail setcred if no authentication
module is in place. We use PAM to establish the dinit session
only when root, so use rootok.

The loginuid however does not make any sense as we will switch
the user immediately afterwards (to the user the service manager
backend runs as).

Also make turnstiled module required, right now it does nothing
and always returns success, but we might want to utilize it at
a later point.

Ref https://github.com/chimera-linux/turnstile/issues/6