populate a login exactly once
There is no need to do it multiple times as long as the service
manager is running, as changed information would mismatch the
info for the running service manager which is not worth handling.

However, do repopulate from passwd when all sessions are dropped
and the service manager has terminated or is to terminate, as then
we can have passwd changes reflect safely without consequences.