implement more advanced session negotiation
Turnstile is now aware of a lot more aspects of the session, such
as VT number, X display, seat, root PID and so on. This will be
necessary later on.

For now things like session state are not handled, they do not
make sense until we have real process tracking.