fix login getting stuck on closed connection
When using blocking sockets, like we are doing in the PAM module,
we will get 0 from recv as EOF; as we are always receiving entire
values of known size, we should block for them and receive at
least some incomplete chunk. If we don't break out on 0, it means
looping forever (as we'll keep receiving 0) and effectively
having the login get stuck.