harden fd_handle_pipe
Prior to this commit, fd_handle_pipe was prone to a DOS attack vector where a user can endlessly spam the readiness pipe. turnstiled would hang if this happens, and eventually run out of memory.

This commit limits the maximum amount of characters the readiness pipe is allowed to read in to PATH_MAX (seems like a reasonable max, but may need adjustment). If the max is exceeded, the ready service is still called as usual, and the value read from the pipe up to that point will be passed. This could lead to issues on the ready service side, but it will allow the user to still log in at least (assuming the ready service isn't critical). Perhaps more should be done to communicate this truncation to the user side, but as of now at least this will prevent the DOS.